+++ internal/domain/incident.go
@@ -10,7 +10,7 @@ type Incident struct {
 	OperatorID string
 	Longitude  float64
 	Latitude   float64
-	Radius     uint8
+	Radius     uint16
 	CreatedAt  time.Time
 	DeletedAt  *time.Time
 	UpdatedAt  *time.Time
